Tests refactor to avoid of the duplication.
