+++ .travis.yml
@@ -31,10 +31,10 @@ matrix:
 #        - OS=ubuntu-18.04
 #        # JOBQUEUE=none is for tests that do not need a cluster to run
 #        - JOBQUEUE=none
-#    - python: "3.6"
-#      env:
-#        - OS=ubuntu-18.04
-#        - JOBQUEUE=slurm
+    - python: "3.6"
+      env:
+        - OS=ubuntu-18.04
+        - JOBQUEUE=slurm
     - python: "3.6"
       env:
         - OS=ubuntu-18.04
