+++ jobqueue_features/tests/test_pbs.py
@@ -6,9 +6,10 @@ from jobqueue_features.clusters_controller import (
     clusters_controller_singleton as controller,
 )
 
+# import logging
+#
+# logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.DEBUG)
 
-import logging
-logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.DEBUG)
 
 class TestPBS(TestCase):
     def setUp(self):
